pyproject.toml
[tool.poetry]
name = "chemistry-ar"
version = "0.1.0"
description = ""
authors = ["Pablo Luque <pablols114@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26"
opencv-contrib-python = "^4.9"
matplotlib = "^3.8"
moderngl = "^5.10.0"
moderngl-window = "^2.4.5"
pyopengl = "^3.1.7"
pywavefront = "^1.3.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
